# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Modular monolith backend (Go + Chi). HTTP handlers live in domain packages (`auth/handler`, `events/handler`, `ticketing/handler`, …) under `internal/app` at base path **`/api/v1`** (отдельного пакета `internal/api/v1` в репозитории нет).\n**Ошибки:** тело `{ \"error\": { \"code\": \"string\", \"message\": \"string\" } }`. Типичные HTTP-коды: **401 Unauthorized** — нет/невалидный JWT (`missing_authorization`, `invalid_token`, `invalid_credentials`, …); **403 Forbidden** — RBAC (`forbidden`, `organizer_request_forbidden`, …); **409 Conflict** — билеты и бизнес-правила (`already_registered`, `capacity_full`, `event_not_approved`, …). Полный список `error.code` — в README.\nДаты событий: **RFC3339**, например `2026-01-01T10:00:00Z`. Ответ `POST /tickets/register`: поля **`qr_hash_hex`** и **`qr_png_base64`**.",
        "title": "Student Event Ticketing Platform API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/analytics/events/stats": {
            "get": {
                "description": "Registration and capacity metrics for an event or aggregated for the caller. **Requires organizer or admin** (students get **403 forbidden**). Organizers see only their events; admins see any event or global aggregates.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Event statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer access token (organizer or admin)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter by event UUID; omit to aggregate events in scope",
                        "name": "event_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.EventStatsResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.EventStatsResponseDTO": {
            "type": "object",
            "properties": {
                "as_of": {
                    "type": "string"
                },
                "event_id": {
                    "type": "string"
                },
                "registered_count": {
                    "type": "integer"
                },
                "registration_timeline": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.RegistrationHour"
                    }
                },
                "remaining_capacity": {
                    "type": "integer"
                },
                "total_capacity": {
                    "type": "integer"
                }
            }
        },
...
        "httpx.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "httpx.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/httpx.ErrorBody"
                }
            }
        },
        "model.RegistrationHour": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "hour": {
                    "type": "string"
                }
            }
        }
    }
}